module_4/requirements.txt
flask==3.1.2
beautifulsoup4==4.13.5
urllib3==2.5.0
requests==2.32.5
psycopg==3.2.3 
pytest==8.4.2
pytest-cov==5.0.0
python-dotenv==1.1.1
sphinx
sphinx-rtd-theme
myst-parser
sphinx-autodoc-typehints